requirements.txt
python-dotenv
selenium
webdriver-manager
pandas
openpyxl
schedule
termcolor